% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gestalt.R
\docType{package}
\name{gestalt-package}
\alias{gestalt}
\alias{gestalt-package}
\title{gestalt: Package Title TBD}
\description{
Package description TBD.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egnha/gestalt}
  \item Report bugs at \url{https://github.com/egnha/gestalt/issues}
}

}
\author{
\strong{Maintainer}: Eugene Ha \email{eha@posteo.de}

}
\keyword{internal}
