% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{let}
\title{Run an Expression in an Ordered Environment}
\usage{
run(..data = parent.frame(), ..expr, ...)

let(..data = parent.frame(), ...)
}
\arguments{
\item{..data}{Environment, list or data frame; if a list or data frame, it is
interpreted as an environment (like the \code{envir} argument of \code{\link[=eval]{eval()}}).}

\item{..expr}{Expression to evaluate (“run”).
\link[rlang:quasiquotation]{Quasiquotation} is supported.}

\item{...}{Named expressions. An expression depends on the preceding ones,
and its name takes precedence over those in \code{..data}. Quasiquotation is
supported.}
}
\value{
\code{run()} returns the evaluation of \code{..expr} in the combined
environment of \code{..data} and \code{...}.

\code{wrt()} returns an environment where the bindings in \code{...} are in
scope, as \link[=delayedAssign]{promises}, as if they were assigned from left
to right in the environment defined by \code{..data}.
}
\description{
Run an Expression in an Ordered Environment
}
