% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrt.R
\name{wrt}
\alias{wrt}
\title{Create a Function with Dependencies}
\usage{
wrt(..., ..env = parent.frame())
}
\arguments{
\item{...}{Function declaration whose body evaluates to a function.
Quasiquotation is supported.}

\item{..env}{Environment in which to create the function.}
}
\value{
Function.
}
\description{
Create a Function with Dependencies
}
\examples{
# A composition with functional dependencies ('b', 'n'):
foo <- wrt(b = 2, n ~ {
  sample \%>>>\% log(base = b) \%>>>\% rep(n)
})

set.seed(1)
foo(1:10, size = 2, n = 3)
#> [1] 1.584963 2.000000 1.584963 2.000000 1.584963 2.000000

set.seed(1)
rep(log(sample(1:10, size = 2), base = 2), 3)
#> [1] 1.584963 2.000000 1.584963 2.000000 1.584963 2.000000

# Because 'wrt()' does the composition upfront, it is faster
# than the equivalent function defined using the magrittr pipe:

library(magrittr)  # Provides the pipe \%>\%

foo_pipe <- function(..., b = 2, n) {
  sample(...) \%>\% log(base = b) \%>\% rep(n)
}

set.seed(1)
foo_pipe(1:10, size = 2, n = 3)
#> [1] 1.584963 2.000000 1.584963 2.000000 1.584963 2.000000

}
