% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{wrt}
\title{Run an Expression in an Ordered Environment}
\usage{
run(..data = parent.frame(), ..expr, ...)

wrt(..env = parent.frame(), ...)
}
\arguments{
\item{..data}{Environment, list or data frame of bindings (cf. \code{envir}
argument of \code{\link[=eval]{eval()}}).}

\item{..expr}{Expression to evaluate (“run”).
\link[rlang:quasiquotation]{Quasiquotation} is supported.}

\item{...}{Named expressions. An expression may depend on the preceding ones.
For \code{run()}, these override those in \code{..data}. Quasiquotation is supported.}

\item{..env}{Environment in which the expressions of \code{...} are resolved.}
}
\value{
\code{run()} returns the evaluation of \code{..expr} in the combined
environment of \code{..data} and \code{...}.

\code{wrt()} returns an environment where the bindings in \code{...} are in
scope, as \link[=delayedAssign]{promises}, as if they were assigned from left
to right in the environment \code{..env}.
}
\description{
Run an Expression in an Ordered Environment
}
