% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\alias{wrt}
\title{Run an Expression in an Ordered Context}
\usage{
run(..data = parent.frame(), ..expr, ...)

wrt(..., ..env = parent.frame())
}
\arguments{
\item{..data}{List or environment of bindings.}

\item{..expr}{Expression to evaluate (“run”). Unquoting is supported.}

\item{...}{Named expressions. An expression may depend on the preceding ones.
Unquoting is supported. For \code{run()}, these override those in \code{..data}.}

\item{..env}{Environment in which the expressions of \code{...} are resolved.}
}
\value{
\code{run()} evaluates \code{..expr} in the combined context defined by
\code{..data} and \code{...}.

\code{wrt()} returns an environment which binds the last expression as a
\link[base:delayedAssign]{promise}. The parent of this environment binds the
penultimate expression as a promise, and so forth, until \code{..env} is reached
as the common ancestor.
}
\description{
Run an Expression in an Ordered Context
}
